function that returns FileSize is now uint64_t to avoid overflow errors on huge files

+++ imfgui/src/model/imfpackageitem.cpp
@@ -26,7 +26,7 @@ std::string IMFPackageItem::GetFileName() const
     return basename(_filename) + ((_type == VIDEO || _type == AUDIO) ? ".mxf" : ".xml");
 }
 
-int IMFPackageItem::GetFileSize() const
+uint64_t IMFPackageItem::GetFileSize() const
 {
     using namespace boost::filesystem;
